# Conflicts:
#	GameEngine/GameLogic/Character/Player/Player.cpp
#	GameEngine/GameLogic/Scene/Trap/Trap.cpp

+++ GameEngine/Source/GameTimer/GameTimer.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Source\LibIncluder.h"
+#include "../LibIncluder.h"
 class GameTimer
 {
 public:
@@ -7,9 +7,9 @@ public:
 	
 
 
-	float GameTime()const; //in seconds
-	float DeltaTime()const; //in seconds
-	float TotalTime() const; // A timer, can be used for in game timers, level time/ bomb explosion time, animation time. anything 
+	double GameTime()const; //in seconds
+	double DeltaTime()const; //in seconds
+	double TotalTime() const; // A timer, can be used for in game timers, level time/ bomb explosion time, animation time. anything 
 
 	void Reset(); //call before message loop
 	void Start(); // call when unpaused
